backend/requirements.txt
fastapi
uvicorn[standard]
dnspython
tldextract
tqdm
email_validator
python-multipart